back-test/bentoml_v1/requirements.txt
aiofiles==22.1.0
aiohttp==3.8.3
aiohttp-cors==0.7.0
aiosignal==1.3.1
albumentations==1.1.0
alembic==1.9.2
altair==4.1.0
anyio==3.6.2
appdirs==1.4.4
argon2-cffi==21.1.0
asgiref==3.4.1
astor==0.8.1
async-timeout==4.0.2
attrs==21.2.0
backcall==0.2.0
backoff==2.2.1
backports.zoneinfo==0.2.1
base58==2.1.1
bentoml==1.0.13
bleach==4.1.0
blinker==1.4
boto3==1.26.57
botocore==1.29.57
build==0.10.0
cachetools==4.2.4
cattrs==22.2.0
Cerberus==1.3.4
certifi==2021.10.8
cffi==1.15.1
chardet==5.1.0
charset-normalizer==2.0.7
circus==0.18.0
click==8.1.2
click-option-group==0.5.5
cloudpickle==2.2.1
configparser==5.3.0
contextlib2==21.6.0
cycler==0.10.0
decorator==4.4.2
deepmerge==1.1.0
defusedxml==0.7.1
Deprecated==1.2.13
docker==5.0.3
efficientnet-pytorch==0.7.1
entrypoints==0.3
exceptiongroup==1.1.0
fastapi==0.68.1
Flask==2.2.2
frozenlist==1.3.3
fs==2.4.16
gitdb==4.0.9
GitPython==3.1.20
googleapis-common-protos==1.58.0
grpcio==1.51.1
gunicorn==20.1.0
h11==0.12.0
humanfriendly==10.0
idna==3.3
imageio==2.9.0
importlib-metadata==6.0.0
importlib-resources==5.10.2
ipykernel==5.5.6
ipython==7.16.1
ipython-genutils==0.2.0
ipywidgets==7.6.5
itsdangerous==2.1.2
jedi==0.18.0
Jinja2==3.0.3
jmespath==1.0.1
joblib==1.1.0
jsonschema==4.0.0
jupyter-client==7.0.6
jupyter-core==4.9.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
kiwisolver==1.3.1
Mako==1.2.4
markdown-it-py==2.1.0
MarkupSafe==2.1.2
matplotlib==3.3.4
mdurl==0.1.2
mistune==0.8.4
multidict==6.0.4
nbclient==0.5.8
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
networkx==2.5.1
notebook==6.4.5
numpy==1.19.5
opencv-python-headless==4.5.1.48
opentelemetry-api==1.14.0
opentelemetry-exporter-otlp-proto-http==1.14.0
opentelemetry-instrumentation==0.35b0
opentelemetry-instrumentation-aiohttp-client==0.35b0
opentelemetry-instrumentation-asgi==0.35b0
opentelemetry-proto==1.14.0
opentelemetry-sdk==1.14.0
opentelemetry-semantic-conventions==0.35b0
opentelemetry-util-http==0.35b0
packaging==21.0
pandas==1.1.5
pandocfilters==1.5.0
parso==0.8.2
pathspec==0.11.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
pip-requirements-parser==32.0.1
pip-tools==6.12.1
prometheus-client==0.12.0
prompt-toolkit==3.0.22
protobuf==3.20.3
psutil==5.9.4
ptyprocess==0.7.0
pyarrow==6.0.0
pycparser==2.21
pydantic==1.7.4
pydeck==0.6.2
Pygments==2.10.0
Pympler==0.9
pynvml==11.4.1
pyparsing==2.4.7
pyproject_hooks==1.0.0
pyrsistent==0.18.0
python-dateutil==2.8.2
python-json-logger==2.0.4
python-multipart==0.0.5
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.1.1
PyYAML==6.0
pyzmq==22.3.0
qudida==0.0.4
requests==2.28.2
rich==13.2.0
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
s3transfer==0.6.0
schema==0.7.5
scikit-image==0.17.2
scikit-learn==0.24.2
scipy==1.5.4
semver==2.13.0
Send2Trash==1.8.0
simple-di==0.1.5
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
SQLAlchemy==1.3.24
SQLAlchemy-Utils==0.36.5
starlette==0.14.2
streamlit==1.17.0
tabulate==0.9.0
terminado==0.12.1
testpath==0.5.0
threadpoolctl==3.0.0
tifffile==2020.9.3
toml==0.10.2
tomli==2.0.1
toolz==0.11.2
torch==1.9.1
torchvision==0.10.1
tornado==6.1
traitlets==4.3.3
typing_extensions==4.4.0
tzdata==2022.7
tzlocal==4.1
urllib3==1.25.11
uvicorn==0.15.0
validators==0.18.2
watchdog==2.1.6
watchfiles==0.18.1
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.4.2
Werkzeug==2.2.2
widgetsnbextension==3.5.2
wrapt==1.14.1
yarl==1.8.2
zipp==3.11.0
